cleanup transporters interface
