+++ runtime/drivers/bigquery/warehouse.go
@@ -274,18 +274,6 @@ func (f *fileIterator) Next() ([]string, error) {
 	return []string{fw.Name()}, nil
 }
 
-// Size implements drivers.FileIterator.
-func (f *fileIterator) Size(unit drivers.ProgressUnit) (int64, bool) {
-	switch unit {
-	case drivers.ProgressUnitRecord:
-		return f.totalRecords, true
-	case drivers.ProgressUnitFile:
-		return 1, true
-	default:
-		return 0, false
-	}
-}
-
 func (f *fileIterator) Format() string {
 	return ""
 }
