+++ runtime/drivers/clickhouse/clickhouse.go
@@ -460,11 +460,6 @@ func (c *connection) AsModelManager(instanceID string) (drivers.ModelManager, bo
 	return c, true
 }
 
-// AsTransporter implements drivers.Connection.
-func (c *connection) AsTransporter(from, to drivers.Handle) (drivers.Transporter, bool) {
-	return nil, false
-}
-
 // AsFileStore implements drivers.Connection.
 func (c *connection) AsFileStore() (drivers.FileStore, bool) {
 	return nil, false
