+++ runtime/drivers/duckdb/model_executor_sqlstore_self.go
@@ -3,11 +3,13 @@ package duckdb
 import (
 	"context"
 	"fmt"
+	"net"
 	"strings"
 
+	"github.com/go-sql-driver/mysql"
 	"github.com/mitchellh/mapstructure"
 	"github.com/rilldata/rill/runtime/drivers"
-	"github.com/rilldata/rill/runtime/drivers/mysql"
+	rillmysql "github.com/rilldata/rill/runtime/drivers/mysql"
 	"github.com/rilldata/rill/runtime/drivers/postgres"
 )
 
@@ -80,7 +82,7 @@ func (e *sqlStoreToSelfExecutor) modelInputProperties(modelName, inputConnector 
 		dsn := inputProps.resolveDSN()
 		if dsn == "" {
 			// may be configured via a connector
-			var config *mysql.ConfigProperties
+			var config *rillmysql.ConfigProperties
 			if err := mapstructure.Decode(inputHandle.Config(), &config); err != nil {
 				return nil, err
 			}
@@ -115,4 +117,45 @@ func (e *sqlStoreToSelfExecutor) modelInputProperties(modelName, inputConnector 
 		return nil, err
 	}
 	return propsMap, nil
+}
+
+// rewriteMySQLDSN rewrites a MySQL DSN to a format that DuckDB expects.
+// DuckDB does not support the URI based DSN format yet. It expects the DSN to be in the form of key=value pairs.
+// This function parses the MySQL URI based DSN and converts it to the key=value format. It only converts the common parameters.
+// For more advanced parameters like SSL configs, the user should manually convert the DSN to the key=value format.
+// If there is an error parsing the DSN, it returns the DSN as is.
+func rewriteMySQLDSN(dsn string) string {
+	cfg, err := mysql.ParseDSN(dsn)
+	if err != nil {
+		// If we can't parse the DSN, just return it as is. May be it is already in the form duckdb expects.
+		return dsn
+	}
+
+	var sb strings.Builder
+
+	if cfg.User != "" {
+		sb.WriteString(fmt.Sprintf("user=%s ", cfg.User))
+	}
+	if cfg.Passwd != "" {
+		sb.WriteString(fmt.Sprintf("password=%s ", cfg.Passwd))
+	}
+	if cfg.DBName != "" {
+		sb.WriteString(fmt.Sprintf("database=%s ", cfg.DBName))
+	}
+	switch cfg.Net {
+	case "unix":
+		sb.WriteString(fmt.Sprintf("socket=%s ", cfg.Addr))
+	case "tcp", "tcp6":
+		host, port, err := net.SplitHostPort(cfg.Addr)
+		if err != nil {
+			return dsn
+		}
+		sb.WriteString(fmt.Sprintf("host=%s ", host))
+		if port != "" {
+			sb.WriteString(fmt.Sprintf("port=%s ", port))
+		}
+	default:
+		return dsn
+	}
+	return sb.String()
 }
