+++ runtime/drivers/https/https.go
@@ -199,11 +199,6 @@ func (c *connection) AsModelManager(instanceID string) (drivers.ModelManager, bo
 	return nil, false
 }
 
-// AsTransporter implements drivers.Connection.
-func (c *connection) AsTransporter(from, to drivers.Handle) (drivers.Transporter, bool) {
-	return nil, false
-}
-
 func (c *connection) AsFileStore() (drivers.FileStore, bool) {
 	return c, true
 }
