+++ runtime/drivers/pinot/pinot.go
@@ -296,10 +296,6 @@ func (c *connection) AsModelManager(instanceID string) (drivers.ModelManager, bo
 	return nil, false
 }
 
-func (c *connection) AsTransporter(from, to drivers.Handle) (drivers.Transporter, bool) {
-	return nil, false
-}
-
 func (c *connection) AsFileStore() (drivers.FileStore, bool) {
 	return nil, false
 }
